+++ src/Analysis/GWASApp/Utils/StateManagement/InitialState.js
@@ -11,6 +11,7 @@ const initialState = {
   finalPopulationSizes: [],
   selectionMode: '',
   messages: [],
+  selectedTeamProject: 'temporary-test-teamproject01', // TODO - remove and leave '' when teamproject selection is ready
 };
 
 export default initialState;
